aws_python_operation/s3_operation/requirements_20250416.txt
boto3==1.37.34
botocore==1.37.34
jmespath==1.0.1
numpy==2.2.4
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
s3transfer==0.11.4
six==1.17.0
tzdata==2025.2
urllib3==2.4.0
